moved tests from FreeStyleJobSpec and MavenJobSpec to JobSpec
