+++ job-dsl-core/src/test/groovy/javaposse/jobdsl/dsl/TestJob.groovy
@@ -1,0 +1,12 @@
+package javaposse.jobdsl.dsl
+
+class TestJob extends Job {
+    TestJob(JobManagement jobManagement) {
+        super(jobManagement)
+    }
+
+    @Override
+    protected String getTemplate() {
+        '<test/>'
+    }
+}
